requirements.txt
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
transformers
datasets
trl
pandas
pyyaml